+++ packages/util-package-finder/HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## 0.3.0 (2019-10-01)
+    * No longer accepts "deprecated" flag
+    * Uses registry.npmjs.com to get list of packages
+
 ## 0.2.0 (2019-07-03)
     * Now returns `date` field in results